refactor: extract provider invoker
